Further fixes and implementation of test relating to privacy

+++ Strava API v3/src/com/danshannon/strava/api/model/SegmentLeaderboardEntry.java
@@ -30,7 +30,7 @@ public class SegmentLeaderboardEntry {
 	private Date startDate;
 	private Date startDateLocal;
 	private Integer activityId;
-	private Integer effortId;
+	private Long effortId;
 	private Integer rank;
 	private String athleteProfile;
 }
